Options package added, refactor clientOptions
